ref(EST-3691-BE): share chunking config
