+++ src/shared/models/knowledge_new.py
@@ -1,7 +1,11 @@
 from typing import Annotated, Any, Literal
 
 from pydantic import Field, field_serializer
-from src.shared.enums.knowledge_new import DocumentStatusEnum, GraphSearchMethodEnum, RAGStrategy
+from src.shared.enums.knowledge_new import (
+    ChunkStrategyEnum,
+    GraphSearchMethodEnum,
+    RAGStrategy
+)
 from src.shared.models.base import ValueObject
 
 __all__ = [
@@ -250,6 +254,15 @@ GraphSearchConfig = Annotated[
 ]
 
 SearchConfig = Annotated[GraphSearchConfig | NaiveSearchConfig, Field(discriminator="rag_strategy")]
+
+
+class ChunkingConfig(ValueObject):
+    """Parameters controlling how a document is chunked."""
+
+    chunk_strategy: ChunkStrategyEnum
+    chunk_size: int
+    chunk_overlap: int
+    extra: dict = Field(default_factory=dict)
 
 
 class PrechunkRequest(ValueObject):
